Migrated user id’s to be Long to match end user id’s

+++ drift/src/main/java/drift/com/drift/adapters/ConversationAdapter.java
@@ -388,7 +388,7 @@ public class ConversationAdapter extends RecyclerView.Adapter<RecyclerView.ViewH
             }
 
 
-            setupForUser(message, UserManager.getInstance().getUserForId(message.authorId.intValue()));
+            setupForUser(message, UserManager.getInstance().getUserForId(message.authorId));
 
         }
